moved the class and interfaces out
