+++ .github/scripts/create-e2e-test-report.ts
@@ -2,6 +2,7 @@ import * as fs from 'fs/promises';
 import humanizeDuration from 'humanize-duration';
 import path from 'path';
 import * as xml2js from 'xml2js';
+import { TestCase, TestRun, TestSuite } from './shared/test-reports';
 import { normalizeTestPath } from './shared/utils';
 
 const XML = {
@@ -43,81 +44,6 @@ function consoleBold(str: string): string {
     .replaceAll('<strong>', '\x1b[1m')
     .replaceAll('</strong>', '\x1b[0m');
 }
-
-class TestRun {
-  testFiles: Record<string, TestFile> = {};
-
-  addTestSuite(testSuite: TestSuite, path: string, allowDuplicates: boolean) {
-    const testFile = this.testFiles[path];
-
-    if (testFile) {
-      const existingSuite = testFile.testSuites.find(
-        (suite) => suite.name === testSuite.name,
-      );
-
-      if (!existingSuite || allowDuplicates) {
-        testFile.testSuites.push(testSuite);
-        testFile.totalTime += testSuite.time;
-        testFile.totalTests += testSuite.tests;
-        testFile.totalPassed += testSuite.passed;
-        testFile.totalFailed += testSuite.failed;
-        testFile.totalSkipped += testSuite.skipped;
-      } else {
-        console.warn(
-          `Test suite ${testSuite.name} already exists in ${path}. Skipping duplicate.`,
-        );
-      }
-    } else {
-      this.testFiles[path] = {
-        path,
-        totalTime: testSuite.time,
-        totalTests: testSuite.tests,
-        totalPassed: testSuite.passed,
-        totalFailed: testSuite.failed,
-        totalSkipped: testSuite.skipped,
-        testSuites: [testSuite],
-      };
-    }
-  }
-}
-
-interface TestFile {
-  path: string;
-  totalTime: number;
-  totalTests: number;
-  totalPassed: number;
-  totalFailed: number;
-  totalSkipped: number;
-  testSuites: TestSuite[];
-}
-
-interface TestSuite {
-  name: string;
-  job: {
-    name: string;
-    id: string;
-  };
-  date: Date;
-  tests: number;
-  passed: number;
-  failed: number;
-  skipped: number;
-  time: number;
-  testCases: TestCase[];
-}
-
-type TestCase =
-  | {
-      name: string;
-      time: number;
-      status: 'passed';
-    }
-  | {
-      name: string;
-      time: number;
-      status: 'failed';
-      error: string;
-    };
 
 async function main() {
   const env = {
